add Bookie::Cell:Navigation
