+++ app/concepts/bookie/view/navigation.slim
@@ -1,0 +1,4 @@
+nav.panel
+  .panel-heading.is-size-6
+    = title
+  == links
