+++ app/concepts/user/notifications/cell/navigation.rb
@@ -1,24 +1,23 @@
-require_dependency 'bookie/cell'
+require_dependency 'bookie/cell/navigation'
 
 class User
   module Notifications
     module Cell
-      class Navigation < Bookie::Cell
+      class Navigation < Bookie::Cell::Navigation
 
         private
 
-        def link_to_friendship_requests
-          # FIXME: ...
-          concept('user/settings/cell/navigation/link', nil, {
+        def title
+          'Notifications'
+        end
+
+        def items
+          [{
+            id: :friendship_requests,
             url: notifications_friendship_requests_path,
             name: 'Friendship requests',
             icon: :organization,
-            active: active == :friendship_requests,
-          })
-        end
-
-        def active
-          options.fetch(:active)
+          }]
         end
       end
     end
