--- app/concepts/user/notifications/view/navigation.slim
@@ -1,4 +1,0 @@
-nav.panel
-  .panel-heading.is-size-6
-    | Notifications
-  = link_to_friendship_requests
