--- app/concepts/user/settings/view/navigation.slim
@@ -1,5 +1,0 @@
-nav.panel
-  .panel-heading.is-size-6
-    | Personal settings
-  = link_to_account
-  = link_to_security
